Adjusted bar chart tooltips
Added "afterTitle" to the callbacks for the tooltips so that "-player games" could be appended to the default numbers representing the x axis values.

Example: "2-player games"

However, there is a issue with this displaying as block rather than inline, such that in the example "-player games" displays on a new line below the "2".

Need to troubleshoot this further.